@@ -1,13 +1,17 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from 'react-dom/client';
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 import {BrowserRouter, Route, Routes} from "react-router-dom";
-import App from "./App";
-import InternalOrdersOverview from "./internal-orders/InternalOrdersOverview";
+import {App} from "./App";
+import {InternalOrdersOverview} from "./modules/InternalOrdersOverview";
 import {SnackbarProvider} from "notistack";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(
+    document.getElementById('root') as HTMLElement
+);
+
+root.render(
     <BrowserRouter>
         <Routes>
             <Route path="/" element={
@@ -31,8 +35,7 @@ ReactDOM.render(
                 />
             </Route>
         </Routes>
-    </BrowserRouter>,
-    document.getElementById("root")
+    </BrowserRouter>
 );
 
 // If you want to start measuring performance in your app, pass a function
