@@ -3,18 +3,22 @@ package de.niko.pcstore.configuration;
 import java.util.Arrays;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
 import org.springframework.core.env.Environment;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.NoOpPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
 
 
-@EnableWebSecurity(debug = true)
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
+@EnableWebSecurity
+public class SecurityConfig {
     private static Logger LOGGER = LoggerFactory.getLogger(SecurityConfig.class);
+
+    public static final String SECURITY_CONFIG_NAME = "App Bearer token";
 
     final Environment environment;
 
@@ -22,12 +26,19 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         this.environment = environment;
     }
 
-    @Override
-    public void configure(WebSecurity web) {
-        web.ignoring().antMatchers("/**");
+//    @Bean
+//    public WebSecurityCustomizer webSecurityCustomizer() {
+//        return (web) -> web.ignoring().antMatchers("/swagger-ui/**");
+//    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+//        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
+        return NoOpPasswordEncoder.getInstance();
     }
 
-    protected void configure(HttpSecurity http) throws Exception {
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         var activeProfiles = environment.getActiveProfiles();
         var activeProfilesList = Arrays.asList(activeProfiles);
 
@@ -35,56 +46,43 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         var isDev = activeProfilesList.contains("dev");
         var isProd = activeProfilesList.contains("prod");
 
-        if (isLocal) {
-            LOGGER.debug("Local environment");
-            http.authorizeRequests().antMatchers("/**").permitAll();
-        } else if (isDev) {
-            LOGGER.debug("Development environment");
-            http.authorizeRequests().anyRequest().authenticated().and().httpBasic();
-//            No auth
-        } else if (isProd) {
-            LOGGER.debug("Production environment");
-        }
+        http
+                .csrf().disable()
+//                .authorizeRequests().antMatchers("/internal**").authenticated().and()
+                .authorizeRequests().mvcMatchers("/swagger-ui/**", "/v3/api-docs/**").permitAll().and()
+                .authorizeRequests().anyRequest().authenticated().and()
+                // httpBasic authentication
+                .httpBasic();
+
+        return http.build();
+
+//        if (isLocal || activeProfilesList.size() == 0) {
+//            LOGGER.info("Local environment");
+////            http.authorizeRequests().antMatchers("/**").permitAll();
+//
+////            http.authorizeRequests()
+////                    .antMatchers("/swagger-ui/**").permitAll()
+////
+////                    .and()
+////                    .authorizeRequests()
+////                   .antMatchers("/internal**").authenticated()
+//////
+//////                    .hasRole("SWAGGER")
+//////
+////                    .and()
+////                    .httpBasic();
+//
+////            http
+////                    .authorizeRequests().antMatchers("/internal-order/**").hasRole("SWAGGER").and().httpBasic();
+//
+//        } else if (isDev) {
+//            LOGGER.info("Development environment");
+//            http.authorizeRequests().anyRequest().authenticated().and().httpBasic();
+////            No auth
+//        } else if (isProd) {
+//            LOGGER.info("Production environment");
+//        }
+//
+//        return http.build();
     }
-
-
-    @Autowired
-    public void globalSecurityConfiguration(AuthenticationManagerBuilder auth) throws Exception {
-        LOGGER.info("globalSecurityConfiguration");
-//        auth.inMemoryAuthentication().withUser("user").password("password").roles("USER");
-//        auth.inMemoryAuthentication().withUser("admin").password("password").roles("USER", "ADMIN");
-    }
-//
-//    @Override
-//    protected void configure(HttpSecurity http) throws Exception {
-////        http.authorizeRequests().antMatchers("/noSecurity").permitAll()
-////                .anyRequest().authenticated()
-////                .and().formLogin().permitAll();
-//        http.authorizeRequests().anyRequest().permitAll().and().csrf().disable();
-//
-//    }
-
-//    @Override
-//    protected void configure(HttpSecurity httpSecurity) throws Exception {
-//        LOGGER.info("configure");
-//
-//        httpSecurity
-//                .authorizeRequests()
-//                .anyRequest().fullyAuthenticated()
-//                .and()
-//                .formLogin();
-//
-////        httpSecurity.authorizeRequests()
-//    }
-
-//    @Override
-//    protected void configure(AuthenticationManagerBuilder authenticationManagerBuilder) throws Exception {
-//        authenticationManagerBuilder
-//                .ldapAuthentication()
-//                .contextSource().url("ldap://localhost:389/")
-//                .managerDn("uid=admin,ou=system").managerPassword("secret")
-//                .and()
-//                .userSearchBase("ou=users,ou=system")
-//                .userSearchFilter("(uid={0})");
-//    }
 }
