@@ -18,19 +18,19 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 public class NewInternalOrderMPDTO {
     @JsonProperty
-    @Schema(description = "Client data", required = true)
+    @Schema(description = "Client data", requiredMode = Schema.RequiredMode.REQUIRED)
     private NewClientDataMPDTO clientData;
 
     @JsonProperty
-    @Schema(description = "Personal computer", required = true)
+    @Schema(description = "Personal computer", requiredMode = Schema.RequiredMode.REQUIRED)
     private NewPersonalComputerMPDTO personalComputer;
 
     @JsonProperty
-    @Schema(description = "Data protection regulation agreement", example = "true", required = true)
+    @Schema(description = "Data protection regulation agreement", example = "true", requiredMode = Schema.RequiredMode.REQUIRED)
     private Boolean privacyPolicy;
 
     @JsonProperty("files")
-    @Schema(description = "files", required = true)
+    @Schema(description = "files", requiredMode = Schema.RequiredMode.REQUIRED)
     private List<NewInternalOrderFileMPDTO> internalOrderFiles;
 
     @Data
@@ -39,31 +39,31 @@ public class NewInternalOrderMPDTO {
     @SuperBuilder
     public static class NewClientDataMPDTO {
         @JsonProperty
-        @Schema(description = "Gender of the customer", example = "male", required = true)
+        @Schema(description = "Gender of the customer", example = "male", requiredMode = Schema.RequiredMode.REQUIRED)
         private Salutation salutation;
 
         @JsonProperty
-        @Schema(description = "Name of the customer", example = "John", required = true)
+        @Schema(description = "Name of the customer", example = "John", requiredMode = Schema.RequiredMode.REQUIRED)
         private String name;
 
         @JsonProperty
-        @Schema(description = "Surname of the customer", example = "Smith", required = true)
+        @Schema(description = "Surname of the customer", example = "Smith", requiredMode = Schema.RequiredMode.REQUIRED)
         private String surname;
 
         @JsonProperty
-        @Schema(description = "Name of the street", example = "Hauptstraße", required = true)
+        @Schema(description = "Name of the street", example = "Hauptstraße", requiredMode = Schema.RequiredMode.REQUIRED)
         private String street;
 
         @JsonProperty
-        @Schema(description = "Number of the house, where the customer lives. Requires only when street specified", example = "110", required = true)
+        @Schema(description = "Number of the house, where the customer lives. Requires only when street specified", example = "110", requiredMode = Schema.RequiredMode.REQUIRED)
         private String houseNumber;
 
         @JsonProperty
-        @Schema(description = "Zip code", example = "90459", required = true)
+        @Schema(description = "Zip code", example = "90459", requiredMode = Schema.RequiredMode.REQUIRED)
         private Integer zip;
 
         @JsonProperty
-        @Schema(description = "Name of the city", example = "Nürnberg", required = true)
+        @Schema(description = "Name of the city", example = "Nürnberg", requiredMode = Schema.RequiredMode.REQUIRED)
         private String city;
 
         @JsonProperty
@@ -112,31 +112,31 @@ public class NewInternalOrderMPDTO {
     @SuperBuilder
     public static class NewPersonalComputerMPDTO {
         @JsonProperty
-        @Schema(description = "Name of the case", example = "MSI MAG Forge 100R", required = true)
+        @Schema(description = "Name of the case", example = "MSI MAG Forge 100R", requiredMode = Schema.RequiredMode.REQUIRED)
         private String computerCase;
 
         @JsonProperty
-        @Schema(description = "Name of the motherboard", example = "Gigabyte B550 Aorus Pro V2", required = true)
+        @Schema(description = "Name of the motherboard", example = "Gigabyte B550 Aorus Pro V2", requiredMode = Schema.RequiredMode.REQUIRED)
         private String motherboard;
 
         @JsonProperty
-        @Schema(description = "Name of the processor", example = "AMD Ryzen 7 5800X", required = true)
+        @Schema(description = "Name of the processor", example = "AMD Ryzen 7 5800X", requiredMode = Schema.RequiredMode.REQUIRED)
         private String processor;
 
         @JsonProperty
-        @Schema(description = "Name of the graphics card", example = "AMD Radeon RX 6700 XT 12 GB", required = true)
+        @Schema(description = "Name of the graphics card", example = "AMD Radeon RX 6700 XT 12 GB", requiredMode = Schema.RequiredMode.REQUIRED)
         private String graphicsCard;
 
         @JsonProperty
-        @Schema(description = "Name of the RAM", example = "32GB Corsair Vengeance LPX DDR4-3000", required = true)
+        @Schema(description = "Name of the RAM", example = "32GB Corsair Vengeance LPX DDR4-3000", requiredMode = Schema.RequiredMode.REQUIRED)
         private String randomAccessMemory;
 
         @JsonProperty
-        @Schema(description = "Name of the storage device", example = "250GB Samsung 870 EVO", required = true)
+        @Schema(description = "Name of the storage device", example = "250GB Samsung 870 EVO", requiredMode = Schema.RequiredMode.REQUIRED)
         private String storageDevice;
 
         @JsonProperty
-        @Schema(description = "Name of the power supply unit", example = "700W - be quiet! Pure power 11", required = true)
+        @Schema(description = "Name of the power supply unit", example = "700W - be quiet! Pure power 11", requiredMode = Schema.RequiredMode.REQUIRED)
         private String powerSupplyUnit;
     }
 
@@ -146,11 +146,11 @@ public class NewInternalOrderMPDTO {
     @SuperBuilder
     public static class NewInternalOrderFileMPDTO {
         @JsonProperty("id")
-        @Schema(description = "Unique UUID", example = "d290f1ee-6c54-4b01-90e6-d701748f0851", required = true)
+        @Schema(description = "Unique UUID", example = "d290f1ee-6c54-4b01-90e6-d701748f0851", requiredMode = Schema.RequiredMode.REQUIRED)
         private String id;
 
         @JsonProperty
-        @Schema(description = "Name of the file", example = "somefile.pdf", required = true)
+        @Schema(description = "Name of the file", example = "somefile.pdf", requiredMode = Schema.RequiredMode.REQUIRED)
         private String name;
 
         @JsonProperty
