@@ -10,7 +10,7 @@ import java.sql.Timestamp;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
-import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.modelmapper.ModelMapper;
@@ -30,7 +30,7 @@ public class MapperConfigurationTest {
     @Test
     @DisplayName("InternalOrderDTO -> InternalOrderEntity")
     public void convertInternalOrderDTO2InternalOrderEntityTest() {
-        Assertions.assertThat(modelMapper).isNotNull();
+        Assertions.assertNotNull(modelMapper);
 
         String internal_order_id = "test-internal-order-id-1";
         Timestamp version = new Timestamp(System.currentTimeMillis());
@@ -43,23 +43,28 @@ public class MapperConfigurationTest {
 
         InternalOrderEntity internalOrderEntity = modelMapper.map(internalOrderDTO, InternalOrderEntity.class);
 
-        Assertions.assertThat(internalOrderEntity).isNotNull();
-        Assertions.assertThat(internalOrderEntity.getId()).isEqualTo(internalOrderDTO.getId()).isEqualTo(internal_order_id);
-        Assertions.assertThat(internalOrderEntity.getVersion()).isEqualTo(internalOrderDTO.getVersion()).isEqualTo(version);
-        Assertions.assertThat(internalOrderEntity.getInternalOrderFileMetadataEntities().size()).isEqualTo(internalOrderDTO.getInternalOrderFiles().size()).isEqualTo(1);
+        Assertions.assertNotNull(internalOrderEntity);
+        Assertions.assertEquals(internal_order_id, internalOrderEntity.getId());
+        Assertions.assertEquals(internal_order_id, internalOrderDTO.getId());
+
+        Assertions.assertEquals(version, internalOrderEntity.getVersion());
+        Assertions.assertEquals(version, internalOrderDTO.getVersion());
+
+        Assertions.assertEquals(1, internalOrderEntity.getInternalOrderFileMetadataEntities().size());
+        Assertions.assertEquals(1, internalOrderDTO.getInternalOrderFiles().size());
 
         internalOrderEntity.getInternalOrderFileMetadataEntities().forEach(internalOrderFileMetadataEntity -> {
-            Assertions.assertThat(internalOrderFileMetadataEntity).isNotNull();
-            Assertions.assertThat(internalOrderFileMetadataEntity.getId()).isEqualTo(internal_order_file_id);
-            Assertions.assertThat(internalOrderFileMetadataEntity.getVersion()).isEqualTo(version);
-            Assertions.assertThat(internalOrderFileMetadataEntity.getName()).isEqualTo(internal_order_file_name);
+            Assertions.assertNotNull(internalOrderFileMetadataEntity);
+            Assertions.assertEquals(internal_order_file_id, internalOrderFileMetadataEntity.getId());
+            Assertions.assertEquals(version, internalOrderFileMetadataEntity.getVersion());
+            Assertions.assertEquals(internal_order_file_name, internalOrderFileMetadataEntity.getName());
         });
     }
 
     @Test
     @DisplayName("InternalOrderEntity -> InternalOrderDTO")
     public void convertInternalOrderEntity2InternalOrderDTOTest() {
-        Assertions.assertThat(modelMapper).isNotNull();
+        Assertions.assertNotNull(modelMapper);
 
         String internal_order_id = "test-internal-order-id-1";
         Timestamp version = new Timestamp(System.currentTimeMillis());
@@ -72,23 +77,30 @@ public class MapperConfigurationTest {
 
         InternalOrderDTO internalOrderDTO = modelMapper.map(internalOrderEntity, InternalOrderDTO.class);
 
-        Assertions.assertThat(internalOrderDTO).isNotNull();
-        Assertions.assertThat(internalOrderDTO.getId()).isEqualTo(internalOrderEntity.getId()).isEqualTo(internal_order_id);
-        Assertions.assertThat(internalOrderDTO.getVersion()).isEqualTo(internalOrderEntity.getVersion()).isEqualTo(version);
-        Assertions.assertThat(internalOrderDTO.getInternalOrderFiles().size()).isEqualTo(internalOrderEntity.getInternalOrderFileMetadataEntities().size()).isEqualTo(1);
+        Assertions.assertNotNull(internalOrderDTO);
+
+        Assertions.assertEquals(internal_order_id, internalOrderDTO.getId());
+        Assertions.assertEquals(internal_order_id, internalOrderEntity.getId());
+
+        Assertions.assertEquals(version, internalOrderDTO.getVersion());
+        Assertions.assertEquals(version, internalOrderEntity.getVersion());
+
+        Assertions.assertEquals(1, internalOrderDTO.getInternalOrderFiles().size());
+        Assertions.assertEquals(1, internalOrderEntity.getInternalOrderFileMetadataEntities().size());
 
         internalOrderDTO.getInternalOrderFiles().forEach(internalOrderFileDTO -> {
-            Assertions.assertThat(internalOrderFileDTO).isNotNull();
-            Assertions.assertThat(internalOrderFileDTO.getId()).isEqualTo(internal_order_file_id);
-            Assertions.assertThat(internalOrderFileDTO.getVersion()).isEqualTo(version);
-            Assertions.assertThat(internalOrderFileDTO.getName()).isEqualTo(internal_order_file_name);
+            Assertions.assertNotNull(internalOrderFileDTO);
+
+            Assertions.assertEquals(internal_order_file_id, internalOrderFileDTO.getId());
+            Assertions.assertEquals(version, internalOrderFileDTO.getVersion());
+            Assertions.assertEquals(internal_order_file_name, internalOrderFileDTO.getName());
         });
     }
 
     @Test
     @DisplayName("PersonalComputerDTO -> PersonalComputerEntity")
     public void convertPersonalComputerDTO2PersonalComputerEntityTest() {
-        Assertions.assertThat(modelMapper).isNotNull();
+        Assertions.assertNotNull(modelMapper);
 
         String id = "test-personal-computer-id-1";
         Timestamp version = new Timestamp(System.currentTimeMillis());
@@ -104,21 +116,30 @@ public class MapperConfigurationTest {
 
         PersonalComputerEntity personalComputerEntity = modelMapper.map(personalComputerDTO, PersonalComputerEntity.class);
 
-        Assertions.assertThat(personalComputerEntity.getId()).isEqualTo(personalComputerDTO.getId()).isEqualTo(id);
-        Assertions.assertThat(personalComputerEntity.getVersion()).isEqualTo(personalComputerDTO.getVersion()).isEqualTo(version);
-        Assertions.assertThat(personalComputerEntity.getComputerCase()).isEqualTo(personalComputerDTO.getComputerCase()).isEqualTo(caseName);
-        Assertions.assertThat(personalComputerEntity.getMotherboard()).isEqualTo(personalComputerDTO.getMotherboard()).isEqualTo(motherboard);
-        Assertions.assertThat(personalComputerEntity.getProcessor()).isEqualTo(personalComputerDTO.getProcessor()).isEqualTo(processor);
-        Assertions.assertThat(personalComputerEntity.getGraphicsCard()).isEqualTo(personalComputerDTO.getGraphicsCard()).isEqualTo(graphicsCard);
-        Assertions.assertThat(personalComputerEntity.getRandomAccessMemory()).isEqualTo(personalComputerDTO.getRandomAccessMemory()).isEqualTo(randomAccessMemory);
-        Assertions.assertThat(personalComputerEntity.getStorageDevice()).isEqualTo(personalComputerDTO.getStorageDevice()).isEqualTo(storageDevice);
-        Assertions.assertThat(personalComputerEntity.getPowerSupplyUnit()).isEqualTo(personalComputerDTO.getPowerSupplyUnit()).isEqualTo(powerSupplyUnit);
+        Assertions.assertEquals(id, personalComputerEntity.getId());
+        Assertions.assertEquals(id, personalComputerDTO.getId());
+        Assertions.assertEquals(version, personalComputerEntity.getVersion());
+        Assertions.assertEquals(version, personalComputerDTO.getVersion());
+        Assertions.assertEquals(caseName, personalComputerEntity.getComputerCase());
+        Assertions.assertEquals(caseName, personalComputerDTO.getComputerCase());
+        Assertions.assertEquals(motherboard, personalComputerEntity.getMotherboard());
+        Assertions.assertEquals(motherboard, personalComputerDTO.getMotherboard());
+        Assertions.assertEquals(processor, personalComputerEntity.getProcessor());
+        Assertions.assertEquals(processor, personalComputerDTO.getProcessor());
+        Assertions.assertEquals(graphicsCard, personalComputerEntity.getGraphicsCard());
+        Assertions.assertEquals(graphicsCard, personalComputerDTO.getGraphicsCard());
+        Assertions.assertEquals(randomAccessMemory, personalComputerEntity.getRandomAccessMemory());
+        Assertions.assertEquals(randomAccessMemory, personalComputerDTO.getRandomAccessMemory());
+        Assertions.assertEquals(storageDevice, personalComputerEntity.getStorageDevice());
+        Assertions.assertEquals(storageDevice, personalComputerDTO.getStorageDevice());
+        Assertions.assertEquals(powerSupplyUnit, personalComputerEntity.getPowerSupplyUnit());
+        Assertions.assertEquals(powerSupplyUnit, personalComputerDTO.getPowerSupplyUnit());
     }
 
     @Test
     @DisplayName("PersonalComputerEntity -> PersonalComputerDTO")
     public void convertPersonalComputerEntity2PersonalComputerDTOTest() {
-        Assertions.assertThat(modelMapper).isNotNull();
+        Assertions.assertNotNull(modelMapper);
 
         String id = "test-personal-computer-id-1";
         Timestamp version = new Timestamp(System.currentTimeMillis());
@@ -134,16 +155,25 @@ public class MapperConfigurationTest {
 
         InternalOrderDTO.PersonalComputerDTO personalComputerDTO = modelMapper.map(personalComputerEntity, InternalOrderDTO.PersonalComputerDTO.class);
 
-        Assertions.assertThat(personalComputerDTO).isNotNull();
-        Assertions.assertThat(personalComputerDTO.getId()).isEqualTo(personalComputerEntity.getId()).isEqualTo(id);
-        Assertions.assertThat(personalComputerDTO.getVersion()).isEqualTo(personalComputerEntity.getVersion()).isEqualTo(version);
-        Assertions.assertThat(personalComputerDTO.getComputerCase()).isEqualTo(personalComputerEntity.getComputerCase()).isEqualTo(caseName);
-        Assertions.assertThat(personalComputerDTO.getMotherboard()).isEqualTo(personalComputerEntity.getMotherboard()).isEqualTo(motherboard);
-        Assertions.assertThat(personalComputerDTO.getProcessor()).isEqualTo(personalComputerEntity.getProcessor()).isEqualTo(processor);
-        Assertions.assertThat(personalComputerDTO.getGraphicsCard()).isEqualTo(personalComputerEntity.getGraphicsCard()).isEqualTo(graphicsCard);
-        Assertions.assertThat(personalComputerDTO.getRandomAccessMemory()).isEqualTo(personalComputerEntity.getRandomAccessMemory()).isEqualTo(randomAccessMemory);
-        Assertions.assertThat(personalComputerDTO.getStorageDevice()).isEqualTo(personalComputerEntity.getStorageDevice()).isEqualTo(storageDevice);
-        Assertions.assertThat(personalComputerDTO.getPowerSupplyUnit()).isEqualTo(personalComputerEntity.getPowerSupplyUnit()).isEqualTo(powerSupplyUnit);
+        Assertions.assertNotNull(personalComputerDTO);
+        Assertions.assertEquals(id, personalComputerDTO.getId());
+        Assertions.assertEquals(id, personalComputerEntity.getId());
+        Assertions.assertEquals(version, personalComputerDTO.getVersion());
+        Assertions.assertEquals(version, personalComputerEntity.getVersion());
+        Assertions.assertEquals(caseName, personalComputerDTO.getComputerCase());
+        Assertions.assertEquals(caseName, personalComputerEntity.getComputerCase());
+        Assertions.assertEquals(motherboard, personalComputerDTO.getMotherboard());
+        Assertions.assertEquals(motherboard, personalComputerEntity.getMotherboard());
+        Assertions.assertEquals(processor, personalComputerDTO.getProcessor());
+        Assertions.assertEquals(processor, personalComputerEntity.getProcessor());
+        Assertions.assertEquals(graphicsCard, personalComputerDTO.getGraphicsCard());
+        Assertions.assertEquals(graphicsCard, personalComputerEntity.getGraphicsCard());
+        Assertions.assertEquals(randomAccessMemory, personalComputerDTO.getRandomAccessMemory());
+        Assertions.assertEquals(randomAccessMemory, personalComputerEntity.getRandomAccessMemory());
+        Assertions.assertEquals(storageDevice, personalComputerDTO.getStorageDevice());
+        Assertions.assertEquals(storageDevice, personalComputerEntity.getStorageDevice());
+        Assertions.assertEquals(powerSupplyUnit, personalComputerDTO.getPowerSupplyUnit());
+        Assertions.assertEquals(powerSupplyUnit, personalComputerEntity.getPowerSupplyUnit());
     }
 
     @Test
@@ -167,12 +197,12 @@ public class MapperConfigurationTest {
 
         InternalOrderShortDTO internalOrderShortDTO = modelMapper.map(internalOrderEntity, InternalOrderShortDTO.class);
 
-        Assertions.assertThat(internalOrderShortDTO).isNotNull();
-        Assertions.assertThat(internalOrderShortDTO.getId()).isNotNull();
-        Assertions.assertThat(internalOrderShortDTO.getVersion()).isNotNull();
-        Assertions.assertThat(internalOrderShortDTO.getClient()).isEqualTo("test surname, test name");
-        Assertions.assertThat(internalOrderShortDTO.getPersonalComputer()).isEqualTo("test cpu, test gpu");
-        Assertions.assertThat(internalOrderShortDTO.getStatus()).isEqualTo(InternalOrderShortDTO.Status.OPEN);
-        Assertions.assertThat(internalOrderShortDTO.getDateOfReceiving()).isEqualTo(dateOfReceiving);
+        Assertions.assertNotNull(internalOrderShortDTO);
+        Assertions.assertEquals(internal_order_id, internalOrderShortDTO.getId());
+        Assertions.assertEquals(version, internalOrderShortDTO.getVersion());
+        Assertions.assertEquals("test surname, test name", internalOrderShortDTO.getClient());
+        Assertions.assertEquals("test cpu, test gpu", internalOrderShortDTO.getPersonalComputer());
+        Assertions.assertEquals(InternalOrderShortDTO.Status.OPEN, internalOrderShortDTO.getStatus());
+        Assertions.assertEquals(dateOfReceiving, internalOrderShortDTO.getDateOfReceiving());
     }
 }
